Add support design library
Move to AppBarLayout to use Toolbar as action bar. This will let the
action bar slide on and off the screen when the content scrolls.
This change required the action bar view to be taken out of the
PanelLayout. In the process NavigationDrawer from the design library was also
used thus making PanelLayout obsolete.