Fix bug where on coming back to start page the action bar would not be visible
If you scroll a web page so that the action bar scrolls off the page and then press the back button to come back to the start page then the action bar would remain off the screen. The only way to bring it back would be to restart the browser or click on one of the tiles on the start page, scroll to bring the action bar back and then come back to the start page.

This commit fixes this issue.